Switch config to latest version of i18next
Switches i18next config to use fs-backend plugin, and adapt config to
new schema.

Removes some config for "missing" functionality, which is only relevant
if a key is used that is not found in the fallback translation.
Previously these missing keys would be written to a file on disk.
AFAICT, this functionality has never been used, so I chose to remove it.

Also removes some config related to the express middleware, which is no
longer relevant as this functionality as been moved to a plugin.