Fix language mismatch flag not set correctly
It appears that previously we were relying on a quirk of i18next, where
the detected language (via header detection), would be set on a
different property to the explicitly set language.

In the latest version this no longer works, so switch to directly using
the language detection to set the showUserOtherLng flag.